♻️ refactor: move llm error classifier to agent runtime
