+++ packages/agent-runtime/src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './llmErrorClassifier';
 export * from './messageSelectors';
 export * from './replay';
 export * from './runtimeRetry';
